export(traceratops): BuildMatrix class to core.build_matrix.py
